refactor: split utilitiy and constants
