+++ src/common/login.js
@@ -1,13 +1,8 @@
 import Cookies from 'js-cookie'
 import { isStorageSupported } from './storage'
-import { brand_name } from './utility'
-import {
-    getDataObjFromCookies,
-    getDataLink,
-    getCookiesFields,
-    getCookiesObject,
-} from './cookies'
+import { getDataObjFromCookies, getDataLink, getCookiesFields, getCookiesObject } from './cookies'
 import { getAppId } from './websocket/config'
+import { brand_name } from 'common/constants'
 
 const Login = (() => {
     const redirectToLogin = () => {
