+++ src/components/hooks/use-livechat.js
@@ -1,12 +1,6 @@
 import React from 'react'
-import {
-    getClientInformation,
-    getDomain,
-    getUTMData,
-    isBrowser,
-    livechat_client_id,
-    livechat_license_id,
-} from 'common/utility'
+import { livechat_client_id, livechat_license_id } from 'common/constants'
+import { getClientInformation, getDomain, getUTMData, isBrowser } from 'common/utility'
 
 export const useLivechat = () => {
     const [is_livechat_interactive, setLiveChatInteractive] = React.useState(false)
@@ -96,12 +90,8 @@ export const useLivechat = () => {
                             residence,
                             first_name,
                             last_name,
-                        } = (client_information) || {};
-                        const {
-                            utm_source,
-                            utm_medium,
-                            utm_campaign,
-                        } = (utm_data) || {};
+                        } = client_information || {}
+                        const { utm_source, utm_medium, utm_campaign } = utm_data || {}
 
                         /* the session variables are sent to CS team dashboard to notify user has logged in
                         and also acts as custom variables to trigger targeted engagement */
@@ -133,8 +123,7 @@ export const useLivechat = () => {
                         if (chat_id) {
                             if (customerSDK) {
                                 customerSDK.on('connected', () => {
-                                    customerSDK?.deactivateChat({ chatId: chat_id }).catch(() => {
-                                    })
+                                    customerSDK?.deactivateChat({ chatId: chat_id }).catch(() => {})
                                 })
                             }
                         }
