+++ src/components/layout/layout.js
@@ -13,7 +13,8 @@ import { LocationProvider } from './location-context'
 import EURedirect, { useModal } from 'components/custom/_eu-redirect-modal.js'
 import CookieBanner from 'components/custom/cookie-banner'
 import { CookieStorage } from 'common/storage'
-import { cfd_warning_height, isBrowser } from 'common/utility'
+import { cfd_warning_height } from 'common/constants'
+import { isBrowser } from 'common/utility'
 import { DerivStore } from 'store'
 import { Localize } from 'components/localization'
 import { Text } from 'components/elements'
