+++ src/components/localization/localized-link.js
@@ -16,9 +16,8 @@ import {
     deriv_bot_app_url,
     smarttrader_url,
     zoho_url,
-    getLocalizedUrl,
-    getDerivAppLocalizedURL,
-} from 'common/utility'
+} from 'common/constants'
+import { getLocalizedUrl, getDerivAppLocalizedURL } from 'common/utility'
 import { DerivStore } from 'store'
 
 const non_localized_links = ['/careers', '/careers/']
@@ -119,7 +118,7 @@ export const LocalizedLink = React.forwardRef(({ to, ...props }, ref) => {
     // Because otherwise this would add a trailing slash
     const { is_default, path, affiliate_lang } = language_config[locale]
     const is_non_localized = non_localized_links.includes(to)
-    const localizedUrl = getLocalizedUrl(path, is_index, to);
+    const localizedUrl = getLocalizedUrl(path, is_index, to)
 
     const path_to = is_default || is_non_localized ? to : localizedUrl
     if (external || external === 'true') {
