+++ src/pages/help-centre/_community.js
@@ -5,7 +5,7 @@ import { Text, Header, QueryImage } from 'components/elements'
 import { LinkButton } from 'components/form'
 import { localize } from 'components/localization'
 import { Container, SectionContainer, Flex } from 'components/containers'
-import { community_url } from 'common/utility'
+import { community_url } from 'common/constants'
 import device from 'themes/device'
 
 const query = graphql`
