+++ src/pages/partners/affiliate-ib/_deriv-affiliate-programme.js
@@ -6,7 +6,7 @@ import { SectionContainer, Container } from 'components/containers'
 import { Header, Text } from 'components/elements/typography'
 import { localize, Localize } from 'components/localization'
 import { LinkButton } from 'components/form'
-import { affiliate_signup_url } from 'common/utility'
+import { affiliate_signup_url } from 'common/constants'
 import device from 'themes/device'
 
 const StyledSection = styled(SectionContainer)`
