+++ src/pages/payment-methods/_dp2p.js
@@ -6,7 +6,8 @@ import { Text, Header, QueryImage } from 'components/elements'
 import { Flex, Show } from 'components/containers'
 import { Button, LinkButton } from 'components/form'
 import { localize, Localize } from 'components/localization'
-import { isBrowser, deriv_dp2p_app_url, dp2p_google_play_url } from 'common/utility'
+import { deriv_dp2p_app_url, dp2p_google_play_url } from 'common/constants'
+import { isBrowser } from 'common/utility'
 import { mobileOSDetect } from 'common/os-detect'
 import Checkmark from 'images/svg/checkmark.svg'
